Review architecture cleanup (#158)
* add md files to lint check and fix

* include note that doc files will be linted

* fix README typo

* updates to Configuration docs

* cleanup and fixes to Financing docs (and matching contract fixes)

* cleanup and fixes to GuildKick docs (and matching contract fixes)

* cleanup Managing docs

* cleanup and fixes to Onboarding docs

* cleanup to Ragequit docs (and matching contract cleanup)

* cleanup to Template doc

* cleanup and fixes to Voting docs

* more cleanup to Configuration docs

* more cleanup to Managing doc

* cleanup to Withdraw doc

* cleanup of HowToCreate

* remove old adapter doc template

* cleanup of DaoRegistry doc

* cleanup of Bank doc

* cleanup of CONTRIBUTING

* run lint:fix

+++ contracts/Migrations.sol
@@ -2,18 +2,18 @@
 pragma solidity ^0.8.0;
 
 contract Migrations {
-  address public owner;
-  uint public last_completed_migration;
+    address public owner;
+    uint256 public last_completed_migration;
 
-  constructor() {
-    owner = msg.sender;
-  }
+    constructor() {
+        owner = msg.sender;
+    }
 
-  modifier restricted() {
-    if (msg.sender == owner) _;
-  }
+    modifier restricted() {
+        if (msg.sender == owner) _;
+    }
 
-  function setCompleted(uint completed) public restricted {
-    last_completed_migration = completed;
-  }
+    function setCompleted(uint256 completed) public restricted {
+        last_completed_migration = completed;
+    }
 }
